Changing project structure to encapsulate functionality that is not reused with the correspondent job. Changing Unit test strategy to avoid IO operations
