+++ impc_etl/workflow/cleaning.py
@@ -6,6 +6,7 @@ class ExperimentCleaner(SparkSubmitTask):
     name = "Experiment_Cleaner"
     app = "impc_etl/jobs/clean/experiment_cleaner.py"
     output_path = luigi.Parameter()
+    entity_type = "experiment"
     dcc_xml_path = luigi.Parameter()
 
     def requires(self):
@@ -18,13 +19,14 @@ class ExperimentCleaner(SparkSubmitTask):
         return ImpcConfig().get_target(output_path)
 
     def app_options(self):
-        return [self.input().path, self.output().path]
+        return [self.input().path, self.entity_type, self.output().path]
 
 
 class LineCleaner(SparkSubmitTask):
     name = "Experiment_Cleaner"
     app = "impc_etl/jobs/clean/line_cleaner.py"
     output_path = luigi.Parameter()
+    entity_type = "line"
     dcc_xml_path = luigi.Parameter()
 
     def requires(self):
@@ -37,7 +39,7 @@ class LineCleaner(SparkSubmitTask):
         return ImpcConfig().get_target(output_path)
 
     def app_options(self):
-        return [self.input().path, self.output().path]
+        return [self.input().path, self.entity_type, self.output().path]
 
 
 class MouseCleaner(SparkSubmitTask):
