+++ tests/integration/test_experiment_normalizer.py
@@ -1,4 +1,4 @@
-from impc_etl.shared.transformations.experiments import _get_closest_weight
+from impc_etl.transformations.experiment_transformations import _get_closest_weight
 from impc_etl.jobs.extract.dcc_extractor import *
 from impc_etl.jobs.clean.experiment_cleaner import *
 from impc_etl.jobs.clean.specimen_cleaner import clean_specimens
