+++ tests/integration/test_static_data_extractor.py
